Extended FoV pit view range and updated jump
Updated the CheckForPit FoV method to extend the view logic to find platforms above the sighted ground. From this pool of the platforms, only the closest platform will be selected as a viable option to jump to. Upgraded the jump methodology with a new jump formula to make up for the excessive height gained from the previous formula in certain cases